% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_services.R
\name{category_service_ops_count}
\alias{category_service_ops_count}
\title{Helper function to count all operations by service and category}
\usage{
category_service_ops_count(in_dir = "../vendor/botocore")
}
\arguments{
\item{in_dir}{Directory containing botocore.}
}
\description{
Helper function to count all operations by service and category
}
