% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{fix_html_span}
\alias{fix_html_span}
\title{Fix HTML span tags}
\usage{
fix_html_span(content, is_rd = FALSE)
}
\arguments{
\item{content}{Character vector (lines of file)}

\item{is_rd}{Logical, TRUE for .Rd files, FALSE for .R files}
}
\value{
Modified character vector
}
\description{
Fix HTML span tags
}
