% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_services.R
\name{list_aws_services}
\alias{list_aws_services}
\title{Helper function to list all aws services and operations from awds js sdk}
\usage{
list_aws_services(in_dir = "../vendor/aws-sdk-js")
}
\arguments{
\item{in_dir}{Directory containing aws js sdk.}
}
\description{
Helper function to list all aws services and operations from awds js sdk
}
