% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_collection.R
\name{make_cran}
\alias{make_cran}
\title{Make packages for CRAN}
\usage{
make_cran(sdk_dir, out_dir, categories = NULL)
}
\arguments{
\item{sdk_dir}{Directory containing the output of \code{make_sdk}.}

\item{out_dir}{Directory to write the packages to.}

\item{categories}{An optional list of categories to use.}
}
\description{
Make packages for CRAN
}
