% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_api.R
\name{make_package_list}
\alias{make_package_list}
\title{Make a list of packages and their antecedant API file names}
\usage{
make_package_list(path)
}
\arguments{
\item{path}{The path to the folder containing the "apis" folder.}
}
\description{
Make a list of packages and their antecedant API file names
}
