% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_single.R
\name{make_single}
\alias{make_single}
\title{Make all single-service packages}
\usage{
make_single(in_dir, sdk_dir, out_dir)
}
\arguments{
\item{in_dir}{Directory containing API files.}

\item{sdk_dir}{Directory containing the output of \code{make_sdk}.}

\item{out_dir}{Directory of the R package.}
}
\description{
Make all single-service packages
}
