% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_build_cran_comments}
\alias{paws_build_cran_comments}
\title{Method to build cran comments}
\usage{
paws_build_cran_comments(
  in_dir = "../cran",
  cache_path = NULL,
  refresh = FALSE
)
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{cache_path}{yaml file created by \code{paws_check_local}}

\item{refresh}{re-write any cran-comments.md that already exists.}
}
\description{
Method to build cran comments
}
