% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_check}
\alias{paws_check}
\title{Check paws sdk}
\usage{
paws_check(in_dir = "../cran", path, keep_notes = FALSE)
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{path}{Path to output paws sdk check results.}

\item{keep_notes}{Include note}
}
\value{
A list of any warnings, errors and notes within paws sdk.
}
\description{
Check paws sdk
}
