% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_check_pkg_size}
\alias{paws_check_pkg_size}
\title{Check paws sdk package size}
\usage{
paws_check_pkg_size(
  in_dir = "../cran",
  threshold = fs::fs_bytes("5MB"),
  pkg_list = list()
)
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{threshold}{\href{https://cran.r-project.org/web/packages/policies.html#:~:text=As\%20a\%20general\%20rule\%2C\%20neither,to\%20a\%20maximum\%20of\%205MB.}{cran package size threshold}
(default 5MB)}

\item{pkg_list}{list of packages check locally, check all packages by default}
}
\value{
A data.table with package status size and percentage.
}
\description{
Check paws sdk package size after being tarball.
}
