% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_check_rhub}
\alias{paws_check_rhub}
\alias{paws_check_rhub_cat}
\alias{paws_check_rhub_sub_cat}
\alias{paws_check_win_devel}
\alias{paws_check_win_devel_cat}
\alias{paws_check_win_devel_sub_cat}
\title{Check paws using rhub}
\usage{
paws_check_rhub(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_rhub_cat(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_rhub_sub_cat(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_win_devel(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_win_devel_cat(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_win_devel_sub_cat(
  in_dir = "../cran",
  pkg_list = list(),
  email = NULL
)
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{pkg_list}{list of packages check through rhub, check all packages by default}

\item{email}{address to notify, defaults to the maintainer address in the package.}
}
\description{
Check paws using rhub
}
