% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_check_url}
\alias{paws_check_url}
\title{Check paws urls}
\usage{
paws_check_url(in_dir = "../cran", path, pkg_list = list())
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{path}{Path to output paws sdk check results.}

\item{pkg_list}{list of packages urls to check, check all packages by default}
}
\description{
Check paws urls
}
