% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_fix_html_span}
\alias{paws_fix_html_span}
\title{Fix html after paws sdk build}
\usage{
paws_fix_html_span(root = "..")
}
\arguments{
\item{root}{Directory containing paws sdk R scripts}
}
\description{
Fix html after paws sdk build
}
