% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_gsub}
\alias{paws_gsub}
\title{Update R script files within paws}
\usage{
paws_gsub(root = "..", before = "", after = "")
}
\arguments{
\item{root}{Directory containing paws sdk R scripts}

\item{before}{character string containing a regular expression to be matched in the given character vector.}

\item{after}{a replacement for matched pattern in sub and gsub.}
}
\description{
Update R script files within paws
}
