% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws_helper_fun.R
\name{paws_new_services}
\alias{paws_new_services}
\title{lists new services from AWS}
\usage{
paws_new_services(in_dir = "../vendor/aws-sdk-js")
}
\arguments{
\item{in_dir}{Directory containing AWS API input files.}
}
\description{
lists new services from AWS
}
