% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_pkg_doc_build}
\alias{paws_pkg_doc_build}
\title{Rebuild paws documentation}
\usage{
paws_pkg_doc_build(in_dir = "../cran", pkg_list = list())
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{pkg_list}{list of packages to release, release all packages by default}
}
\description{
Rebuild paws documentation
}
