% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_post_build_format}
\alias{paws_post_build_format}
\title{Optimized single-pass post-build formatting}
\usage{
paws_post_build_format(root = "..", patterns_before, patterns_after)
}
\arguments{
\item{root}{Directory containing paws sdk R scripts}

\item{patterns_before}{Patterns for gsub}

\item{patterns_after}{Replacements for gsub}
}
\description{
Process files through transformation pipeline with single read/write
}
