% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_release}
\alias{paws_release}
\alias{paws_release_sub_cat}
\alias{paws_release_cat}
\title{Method to release paws sdk to cran}
\usage{
paws_release_sub_cat(in_dir = "../cran", pkg_list = list())

paws_release_cat(in_dir = "../cran", pkg_list = list())

paws_release(in_dir = "../cran", pkg_list = list())
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{pkg_list}{list of packages to release, release all packages by default}
}
\description{
Method to release paws sdk to cran
}
