% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws_helper_fun.R
\name{paws_service_param_fn}
\alias{paws_service_param_fn}
\title{Create paws serice helper functions}
\usage{
paws_service_param_fn(dir = "../paws.common")
}
\arguments{
\item{dir}{directory of paws.common}
}
\description{
Create paws serice helper functions
}
