% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_unescape_latex_post_build}
\alias{paws_unescape_latex_post_build}
\title{Unescape any latex post paws sdk build}
\usage{
paws_unescape_latex_post_build(
  root = "..",
  special_characters = c("#", "$", "_")
)
}
\arguments{
\item{root}{Directory containing paws sdk R scripts}

\item{special_characters}{character to be removed}
}
\description{
Unescape any latex post paws sdk build
}
