% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_update_url}
\alias{paws_update_url}
\title{Update paws urls}
\usage{
paws_update_url(in_dir = "../cran", path, pkg_list = list())
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{path}{Path to output paws sdk update results.}

\item{pkg_list}{list of packages urls to update, update all packages by default}
}
\description{
update paws urls
}
