% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_update_version}
\alias{paws_update_version}
\title{Update paws version}
\usage{
paws_update_version(dir = "../cran", version)
}
\arguments{
\item{dir}{Directory containing paws sdk packages.}

\item{version}{Version to set paws sdk.}
}
\description{
Update paws version
}
