% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{process_file_single_pass}
\alias{process_file_single_pass}
\title{Process single file through transformation pipeline}
\usage{
process_file_single_pass(file, patterns_before, patterns_after, is_rd = FALSE)
}
\arguments{
\item{file}{Path to file}

\item{patterns_before}{Patterns for gsub}

\item{patterns_after}{Replacements for gsub}

\item{is_rd}{Is this an .Rd file?}
}
\value{
TRUE if file was modified, FALSE otherwise
}
\description{
Process single file through transformation pipeline
}
