% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_utils.R
\name{read_api}
\alias{read_api}
\title{Read API definition files given a name, version, and directory}
\usage{
read_api(name, version, path)
}
\description{
Read API definition files given a name, version, and directory
}
