% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_utils.R
\name{read_region_config}
\alias{read_region_config}
\title{Read the region config file, containing special endpoint information.}
\usage{
read_region_config(path)
}
\description{
Read the region config file, containing special endpoint information.
}
