% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run all tests in a package in another R session}
\usage{
run_tests(path, timeout = 180)
}
\arguments{
\item{path}{The path to the package source.}

\item{timeout}{How long it should be allowed to run.}
}
\description{
Run all tests in a package in another R session
}
