% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{unescape_latex}
\alias{unescape_latex}
\title{Unescape LaTeX special characters}
\usage{
unescape_latex(content, special_characters = c("#", "$", "_"))
}
\arguments{
\item{content}{Character vector (lines of file)}

\item{special_characters}{Characters to unescape (default: #, $, _)}
}
\value{
Modified character vector
}
\description{
Unescape LaTeX special characters
}
