% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{make_cran}
\alias{make_cran}
\title{Make category-level packages for CRAN}
\usage{
make_cran(sdk_dir, out_dir)
}
\arguments{
\item{sdk_dir}{Directory containing the output of \code{make_sdk}.}

\item{out_dir}{Directory to write the packages to.}
}
\description{
Make category-level packages for CRAN
}
