% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make an R package for a given API}
\usage{
make_package(api_name, in_dir, out_dir)
}
\arguments{
\item{in_dir}{Directory containing API files.}

\item{out_dir}{Directory to}

\item{api}{Name of the API to make a package for.}
}
\description{
Make an R package for a given API
}
