% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{make_package}
\alias{make_package}
\title{Make an R package for a given API}
\usage{
make_package(api_info, region_config, path, fields = NULL,
  overwrite = FALSE)
}
\arguments{
\item{region_config}{AWS region config, from \link{read_region_config}.}

\item{path}{Path to the directory where the package should be created.}

\item{fields}{List of fields to include in the DESCRIPTION.}

\item{overwrite}{Overwite existing packages?}

\item{api}{AWS API list object, from \link{read_api}.}
}
\description{
Make an R package for a given API
}
