% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sdk.R
\name{make_sdk}
\alias{make_sdk}
\title{Make the AWS SDK R Package}
\usage{
make_sdk(in_dir, out_dir)
}
\arguments{
\item{in_dir}{Directory containing API files.}

\item{out_dir}{Directory of the R package.}
}
\description{
Make the AWS SDK R Package
}
