% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_api.R
\name{make_sdk_for_api}
\alias{make_sdk_for_api}
\title{Add the SDK for a given API}
\usage{
make_sdk_for_api(api_name, in_dir)
}
\arguments{
\item{api_name}{Name of the API to make a package for.}

\item{in_dir}{Directory containing API files.}

\item{out_dir}{Directory of the R package.}
}
\description{
Add the SDK for a given API
}
\keyword{internal}
