% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk_helper.R
\name{paws_check_rhub}
\alias{paws_check_rhub}
\alias{paws_check_rhub_cat}
\alias{paws_check_rhub_sub_cat}
\alias{paws_rhub_action_check}
\alias{paws_check_win_devel}
\alias{paws_check_win_devel_cat}
\alias{paws_check_win_devel_sub_cat}
\title{Check paws using rhub}
\usage{
paws_check_rhub(
  in_dir = "../cran",
  pkg_list = list(),
  platforms = c("linux", "macos", "macos-arm64", "windows")
)

paws_check_rhub_cat(
  in_dir = "../cran",
  pkg_list = list(),
  platforms = c("linux", "macos", "macos-arm64", "windows")
)

paws_check_rhub_sub_cat(
  in_dir = "../cran",
  pkg_list = list(),
  platforms = c("linux", "macos", "macos-arm64", "windows")
)

paws_rhub_action_check(
  packages = "paws.storage",
  platforms = c("linux", "macos", "macos-arm64", "windows")
)

paws_check_win_devel(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_win_devel_cat(in_dir = "../cran", pkg_list = list(), email = NULL)

paws_check_win_devel_sub_cat(
  in_dir = "../cran",
  pkg_list = list(),
  email = NULL
)
}
\arguments{
\item{in_dir}{Directory containing paws sdk packages.}

\item{pkg_list}{list of packages check through rhub, check all packages by default}

\item{platforms}{Platforms to use, a character vector. Use NULL to select from a list in interactive sessions. See \code{rhub::rhub_platforms()}.}
}
\description{
Check paws using rhub
}
